+++ backend/src/compiler/compilers/JDoodleCompilerImpl.ts
@@ -1,5 +1,5 @@
-import { CompilerRequest, CompilerResponse } from "../../types/compiler";
-import { OutputParser } from "../../utils/OutputParser";
+import { CompilerRequest, CompilerResponse } from "../types";
+import { OutputParser } from "../utils/OutputParser";
 import { Compiler } from "./compilerFactory";
 
 // https://www.jdoodle.com/compiler-api/
